refactor(api): move resource initialization into api crate
Move database, Vault, credential, certificate, KMS, and secret-import
setup into the api composition crate. Keep api-core focused on API
implementation and domain logic.
